+++ src/main/java/org/hyperonline/hyperlib/controller/HYPER_PWMSparkMax.java
@@ -2,6 +2,7 @@ package org.hyperonline.hyperlib.controller;
 
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import edu.wpi.first.wpilibj.motorcontrol.PWMSparkMax;
+import org.hyperonline.hyperlib.controller.meta.RawController;
 
 public class HYPER_PWMSparkMax extends PWMSparkMax implements RawController {
     /**
