+++ src/main/java/org/hyperonline/hyperlib/controller/meta/groups/ControllerGroup.java
@@ -1,8 +1,8 @@
-package org.hyperonline.hyperlib.controller.groups;
+package org.hyperonline.hyperlib.controller.meta.groups;
 
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import edu.wpi.first.util.sendable.SendableBuilder;
-import org.hyperonline.hyperlib.controller.MetaController;
+import org.hyperonline.hyperlib.controller.meta.MetaController;
 import org.hyperonline.hyperlib.controller.SendableMotorController;
 
 /**
